Remove some pub functions from core
